added type to shared
